Added my initial files
This module was created by using behave and not cucumber. Andrew Knight's videos are outdated and were presented with many errors. I found a very useful behave introduction on this website and I was able to implement it to my code. Here is the link to the site: https://blog.visionify.ai/bdd-quickstart-with-python-4cf366cfc11c 

Travis CI is still not properly functioning.